Modified price field from Integer to Float type

+++ warehouse_product/models.py
@@ -6,7 +6,8 @@ from product.models import Product
 class WarehouseProduct(models.Model):
     warehouse = models.ForeignKey(Warehouse, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    price = models.IntegerField()
+    # price = models.IntegerField()
+    price = models.FloatField()
     stock = models.IntegerField()
     discount_rate = models.IntegerField(default=0)
     created_date = models.DateTimeField(auto_now_add=True)
